// src/types/types.ts
export type DataType = {
  copyright: string;
  date: string;
  explanation: string;
  hdurl: string;
  media_type: string;
  service_version: string;
  title: string;
  url: string;
};

export type CustomDateType = {
  date: string;
  startDate: string;
  endDate: string;
};

export type ParamsType = {
  api_key: string;
  date: string;
  start_date: string;
  end_date: string;
};